# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/wallet/public/api/"
        }
    ],
    "paths": {
        "/V1/payments": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "index Payment",
                "description": "index Payment",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful operation"
                    },
                    "202": {
                        "description": "Successful operation"
                    },
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "store Payment",
                "description": "store Payment",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful operation"
                    },
                    "202": {
                        "description": "Successful operation"
                    },
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/V1/payments/{id}": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "show Payment",
                "description": "show Payment",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "payment id",
                        "required": true,
                        "example": "1234567890qwer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful operation"
                    },
                    "202": {
                        "description": "Successful operation"
                    },
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Payment"
                ],
                "summary": "Delete Payment",
                "description": "Delete  payment",
                "operationId": "DeletePayment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment id",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Payment Successfuly Removed"
                    },
                    "403": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/V1/payments/{id}/reject": {
            "patch": {
                "tags": [
                    "Payment"
                ],
                "summary": "reject Payment",
                "description": "reject Payment",
                "operationId": "reject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "payment id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful operation"
                    },
                    "202": {
                        "description": "Successful operation"
                    },
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/V1/payments/{id}/approve": {
            "patch": {
                "tags": [
                    "Payment"
                ],
                "summary": "approve Payment",
                "description": "approve Payment",
                "operationId": "approve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "payment id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful operation"
                    },
                    "202": {
                        "description": "Successful operation"
                    },
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    }
}